android: small ui tweaks
