+++ android/app/src/main/java/me/kavishdevar/librepods/composables/AdaptiveStrengthSlider.kt
@@ -20,22 +20,12 @@
 
 package me.kavishdevar.librepods.composables
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Slider
-import androidx.compose.material3.SliderDefaults
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.DisposableEffect
@@ -43,19 +33,19 @@ import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import me.kavishdevar.librepods.R
 import me.kavishdevar.librepods.services.ServiceManager
 import me.kavishdevar.librepods.utils.AACPManager
 import kotlin.io.encoding.ExperimentalEncodingApi
 import kotlin.math.roundToInt
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun AdaptiveStrengthSlider() {
     val sliderValue = remember { mutableFloatStateOf(0f) }
@@ -100,80 +90,20 @@ fun AdaptiveStrengthSlider() {
 
     Column(
         modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 8.dp),
+            .fillMaxWidth(),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        Slider(
-            value = sliderValue.floatValue,
+        StyledSlider(
+            mutableFloatState = sliderValue,
             onValueChange = {
                 sliderValue.floatValue = snapIfClose(it, listOf(0f, 50f, 100f))
             },
             valueRange = 0f..100f,
-            onValueChangeFinished = {
-                sliderValue.floatValue = snapIfClose(sliderValue.floatValue.roundToInt().toFloat(), listOf(0f, 50f, 100f))
-                service.aacpManager.sendControlCommand(
-                    identifier = AACPManager.Companion.ControlCommandIdentifiers.AUTO_ANC_STRENGTH.value,
-                    value = (100 - sliderValue.floatValue).toInt()
-                )
-            },
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(36.dp),
-            colors = SliderDefaults.colors(
-                thumbColor = thumbColor,
-                inactiveTrackColor = trackColor
-            ),
-            thumb = {
-                Box(
-                    modifier = Modifier
-                        .size(24.dp)
-                        .shadow(4.dp, CircleShape)
-                        .background(thumbColor, CircleShape)
-                )
-            },
-            track = {
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(12.dp),
-                    contentAlignment = Alignment.CenterStart
-                )
-                {
-                    Box(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(4.dp)
-                            .background(trackColor, RoundedCornerShape(4.dp))
-                    )
-                }
-
-            }
+            snapPoints = listOf(0f, 50f, 100f),
+            startLabel = stringResource(R.string.less_noise),
+            endLabel = stringResource(R.string.more_noise),
+            independent = false
         )
-
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-            Text(
-                text = "Less Noise",
-                style = TextStyle(
-                    fontSize = 14.sp,
-                    fontWeight = FontWeight.Light,
-                    color = labelTextColor
-                ),
-                modifier = Modifier.padding(start = 4.dp)
-            )
-            Text(
-                text = "More Noise",
-                style = TextStyle(
-                    fontSize = 14.sp,
-                    fontWeight = FontWeight.Light,
-                    color = labelTextColor
-                ),
-                modifier = Modifier.padding(end = 4.dp)
-            )
-        }
     }
 }
