+++ android/app/src/main/java/me/kavishdevar/librepods/screens/HearingAidAdjustmentsScreen.kt
@@ -343,17 +343,8 @@ fun HearingAidAdjustmentsScreen(@Suppress("unused") navController: NavController
                 }
             }
 
-            Text(
-                text = stringResource(R.string.amplification).uppercase(),
-                style = TextStyle(
-                    fontSize = 14.sp,
-                    fontWeight = FontWeight.Light,
-                    color = textColor.copy(alpha = 0.6f),
-                    fontFamily = FontFamily(Font(R.font.sf_pro))
-                ),
-                modifier = Modifier.padding(8.dp, bottom = 0.dp)
-            )
             StyledSlider(
+                label = stringResource(R.string.amplification).uppercase(),
                 valueRange = -1f..1f,
                 mutableFloatState = amplificationSliderValue,
                 onValueChange = {
@@ -374,17 +365,8 @@ fun HearingAidAdjustmentsScreen(@Suppress("unused") navController: NavController
                 description = stringResource(R.string.swipe_amplification_description)
             )
 
-            Text(
-                text = stringResource(R.string.balance).uppercase(),
-                style = TextStyle(
-                    fontSize = 14.sp,
-                    fontWeight = FontWeight.Light,
-                    color = textColor.copy(alpha = 0.6f),
-                    fontFamily = FontFamily(Font(R.font.sf_pro))
-                ),
-                modifier = Modifier.padding(8.dp, bottom = 0.dp)
-            )
             StyledSlider(
+                label = stringResource(R.string.balance).uppercase(),
                 valueRange = -1f..1f,
                 mutableFloatState = balanceSliderValue,
                 onValueChange = {
@@ -396,17 +378,8 @@ fun HearingAidAdjustmentsScreen(@Suppress("unused") navController: NavController
                 independent = true,
             )
 
-            Text(
-                text = stringResource(R.string.tone).uppercase(),
-                style = TextStyle(
-                    fontSize = 14.sp,
-                    fontWeight = FontWeight.Light,
-                    color = textColor.copy(alpha = 0.6f),
-                    fontFamily = FontFamily(Font(R.font.sf_pro))
-                ),
-                modifier = Modifier.padding(8.dp, bottom = 0.dp)
-            )
             StyledSlider(
+                label = stringResource(R.string.tone).uppercase(),
                 valueRange = -1f..1f,
                 mutableFloatState = toneSliderValue,
                 onValueChange = {
@@ -417,18 +390,8 @@ fun HearingAidAdjustmentsScreen(@Suppress("unused") navController: NavController
                 independent = true,
             )
 
-            Text(
-                text = stringResource(R.string.ambient_noise_reduction).uppercase(),
-                style = TextStyle(
-                    fontSize = 14.sp,
-                    fontWeight = FontWeight.Light,
-                    color = textColor.copy(alpha = 0.6f),
-                    fontFamily = FontFamily(Font(R.font.sf_pro))
-                ),
-                modifier = Modifier.padding(8.dp, bottom = 0.dp)
-            )
-
             StyledSlider(
+                label = stringResource(R.string.ambient_noise_reduction).uppercase(),
                 valueRange = 0f..1f,
                 mutableFloatState = ambientNoiseReductionSliderValue,
                 onValueChange = {
